Improve exec.Cmd interface
