+++ providers/kube/exec.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"strings"
 
+	"go.ligato.io/vpp-probe/pkg/exec"
 	"go.ligato.io/vpp-probe/providers/kube/client"
 )
 
@@ -21,16 +22,19 @@ type command struct {
 	pod *client.Pod
 }
 
-func (c *command) SetStdin(in io.Reader) {
+func (c *command) SetStdin(in io.Reader) exec.Cmd {
 	c.Stdin = in
+	return c
 }
 
-func (c *command) SetStdout(out io.Writer) {
+func (c *command) SetStdout(out io.Writer) exec.Cmd {
 	c.Stdout = out
+	return c
 }
 
-func (c *command) SetStderr(out io.Writer) {
+func (c *command) SetStderr(out io.Writer) exec.Cmd {
 	c.Stderr = out
+	return c
 }
 
 func (c *command) Output() ([]byte, error) {
